+++ libft/ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: tpetit <tpetit@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/16 16:56:50 by tpetit            #+#    #+#             */
-/*   Updated: 2020/11/16 20:11:06 by tpetit           ###   ########.fr       */
+/*   Updated: 2020/11/18 11:20:28 by tpetit           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*ns1;
 	unsigned char	*ns2;
